refactor: minimize 로직을 hook으로 분리
